refacto conftest
