+++ api/tests/integration/endpoints/conftest.py
@@ -1,5 +1,4 @@
 import pytest_asyncio
-import redis.asyncio as redis
 
 
 @pytest_asyncio.fixture(scope="function", autouse=True)
@@ -8,14 +7,3 @@ async def _restore_dependency_overrides(app):
     yield
     app.dependency_overrides.clear()
     app.dependency_overrides.update(snapshot)
-
-
-@pytest_asyncio.fixture(scope="function", autouse=True)
-async def _reset_redis_between_tests(test_redis_pool):
-    client = redis.Redis(connection_pool=test_redis_pool)
-    try:
-        await client.flushdb()
-        yield
-        await client.flushdb()
-    finally:
-        await client.aclose()
